deploy: split out tunnel deploy and setup files
