+++ deploy/06_asset/17_set_asset_erc1155_tunnel.ts
@@ -1,0 +1,51 @@
+import {DeployFunction} from 'hardhat-deploy/types';
+import {HardhatRuntimeEnvironment} from 'hardhat/types';
+import {skipUnlessTestnet} from '../../utils/network';
+
+const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
+  const {deployments, getNamedAccounts} = hre;
+  const {execute, read, catchUnknownSigner} = deployments;
+  const {deployer} = await getNamedAccounts();
+
+  const AssetERC1155Tunnel = await deployments.get('AssetERC1155Tunnel');
+  const currentAdmin = await read('Asset', 'getAdmin');
+  await catchUnknownSigner(
+    execute(
+      'Asset',
+      {from: currentAdmin, log: true},
+      'setPredicate',
+      AssetERC1155Tunnel.address
+    )
+  );
+
+  const PolygonAssetERC1155Tunnel = await hre.companionNetworks[
+    'l2'
+  ].deployments.getOrNull('PolygonAssetERC1155Tunnel');
+
+  // get deployer on l2
+  const {deployer: deployerOnL2} = await hre.companionNetworks[
+    'l2'
+  ].getNamedAccounts();
+
+  if (PolygonAssetERC1155Tunnel) {
+    await hre.companionNetworks['l2'].deployments.execute(
+      'PolygonAssetERC1155Tunnel',
+      {from: deployerOnL2},
+      'setFxRootTunnel',
+      AssetERC1155Tunnel.address
+    );
+    await deployments.execute(
+      'AssetERC1155Tunnel',
+      {from: deployer},
+      'setFxChildTunnel',
+      PolygonAssetERC1155Tunnel.address
+    );
+  }
+};
+
+export default func;
+func.tags = ['AssetERC1155Tunnel_setup', 'L1'];
+func.runAtTheEnd = true;
+func.dependencies = ['AssetERC1155Tunnel_deploy'];
+func.skip = skipUnlessTestnet;
+func.runAtTheEnd = true;
