+++ deploy_polygon/02_asset/06_deploy_polygon_asset_erc721_tunnel.ts
@@ -5,7 +5,7 @@ import {skipUnlessTestnet} from '../../utils/network';
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const {deployments, getNamedAccounts} = hre;
-  const {deploy, read} = deployments;
+  const {deploy} = deployments;
   const {deployer} = await getNamedAccounts();
 
   const TRUSTED_FORWARDER = await deployments.get('TRUSTED_FORWARDER');
@@ -14,7 +14,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
 
   const MAX_TRANSFER_LIMIT = 20;
 
-  const PolygonAssetERC721Tunnel = await deploy('PolygonAssetERC721Tunnel', {
+  await deploy('PolygonAssetERC721Tunnel', {
     from: deployer,
     contract: 'PolygonAssetERC721Tunnel',
     args: [
@@ -26,42 +26,6 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
     log: true,
     skipIfAlreadyDeployed: true,
   });
-
-  const AssetERC721Tunnel = await hre.companionNetworks[
-    'l1'
-  ].deployments.getOrNull('AssetERC721Tunnel');
-  // get deployer on l1
-  const {deployer: deployerOnL1} = await hre.companionNetworks[
-    'l1'
-  ].getNamedAccounts();
-
-  // Note: if redeploy, delete tunnels
-
-  if (AssetERC721Tunnel) {
-    const fxChildTunnel = await hre.companionNetworks['l1'].deployments.read(
-      'AssetERC721Tunnel',
-      'fxChildTunnel'
-    );
-    const fxRootTunnel = await read('PolygonAssetERC721Tunnel', 'fxRootTunnel');
-
-    if (fxRootTunnel == constants.AddressZero) {
-      await deployments.execute(
-        'PolygonAssetERC721Tunnel',
-        {from: deployer, log: true},
-        'setFxRootTunnel',
-        AssetERC721Tunnel.address
-      );
-    }
-
-    if (fxChildTunnel == constants.AddressZero) {
-      await hre.companionNetworks['l1'].deployments.execute(
-        'AssetERC721Tunnel',
-        {from: deployerOnL1, log: true},
-        'setFxChildTunnel',
-        PolygonAssetERC721Tunnel.address
-      );
-    }
-  }
 };
 
 export default func;
