+++ deploy_polygon/02_asset/09_deploy_polygon_asset_erc1155_tunnel.ts
@@ -5,7 +5,7 @@ import {skipUnlessTestnet} from '../../utils/network';
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const {deployments, getNamedAccounts} = hre;
-  const {deploy, read} = deployments;
+  const {deploy} = deployments;
   const {deployer} = await getNamedAccounts();
 
   const PolygonAssetERC1155 = await deployments.get('PolygonAssetERC1155');
@@ -14,7 +14,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
 
   const MAX_TRANSFER_LIMIT = 20;
 
-  const PolygonAssetERC1155Tunnel = await deploy('PolygonAssetERC1155Tunnel', {
+  await deploy('PolygonAssetERC1155Tunnel', {
     from: deployer,
     contract: 'PolygonAssetERC1155Tunnel',
     args: [
@@ -26,45 +26,6 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
     log: true,
     skipIfAlreadyDeployed: true,
   });
-
-  const AssetERC1155Tunnel = await hre.companionNetworks[
-    'l1'
-  ].deployments.getOrNull('AssetERC1155Tunnel');
-  // get deployer on l1
-  const {deployer: deployerOnL1} = await hre.companionNetworks[
-    'l1'
-  ].getNamedAccounts();
-
-  // Note: if redeploy, delete tunnels
-
-  if (AssetERC1155Tunnel) {
-    const fxChildTunnel = await hre.companionNetworks['l1'].deployments.read(
-      'AssetERC1155Tunnel',
-      'fxChildTunnel'
-    );
-    const fxRootTunnel = await read(
-      'PolygonAssetERC1155Tunnel',
-      'fxRootTunnel'
-    );
-
-    if (fxRootTunnel == constants.AddressZero) {
-      await deployments.execute(
-        'PolygonAssetERC1155Tunnel',
-        {from: deployer, log: true},
-        'setFxRootTunnel',
-        AssetERC1155Tunnel.address
-      );
-    }
-
-    if (fxChildTunnel == constants.AddressZero) {
-      await hre.companionNetworks['l1'].deployments.execute(
-        'AssetERC1155Tunnel',
-        {from: deployerOnL1, log: true},
-        'setFxChildTunnel',
-        PolygonAssetERC1155Tunnel.address
-      );
-    }
-  }
 };
 
 export default func;
